Move the code from old file to the new files
